+++ MQTTPacket/src/MQTTSubscribeServer.c
@@ -50,7 +50,7 @@ int32_t MQTTDeserialize_subscribe(unsigned char* dup, unsigned short* packetid, 
 	unsigned char* curdata = buf;
 	unsigned char* enddata = NULL;
 	int32_t rc = MQTTPACKET_READ_ERROR;
-	size_t mylen = 0;
+	uint32_t mylen = 0;
 
 	FUNC_ENTRY;
 	header = readChar(&curdata);
